chore: I swear to all things holy
* This doesn't break anything, but I'd rather not have compiler warnings argh

+++ plip-sdl/Gui.cpp
@@ -66,7 +66,7 @@ void Gui::DrawBreakpointControls() {
                     ImGui::TableSetBgColor(ImGuiTableBgTarget_CellBg, bpHitHighlight);
                 }
 
-                ImGui::Text("0x%.8llX", static_cast<uint64_t>(bp));
+                ImGui::Text("0x%.8llX", static_cast<unsigned long long>(bp));
 
                 ImGui::TableNextColumn();
                 ImGui::PushID(static_cast<int>(bp));  // TODO: Revisit this if we (somehow) implement systems with 32-bit addressing.
@@ -105,22 +105,22 @@ void Gui::DrawCoreDebugInfo() {
                         break;
 
                     case Plip::DebugValueType::Int8:
-                        ImGui::Text("0x%.2llX (%llu)", static_cast<uint64_t>(value.ValueInt), static_cast<uint64_t>(value.ValueInt));
+                        ImGui::Text("0x%.2llX (%llu)", static_cast<unsigned long long>(value.ValueInt), static_cast<unsigned long long>(value.ValueInt));
                         break;
 
                     case Plip::DebugValueType::Int16Le:
                     case Plip::DebugValueType::Int16Be:
-                        ImGui::Text("0x%.4llX (%llu)", static_cast<uint64_t>(value.ValueInt), static_cast<uint64_t>(value.ValueInt));
+                        ImGui::Text("0x%.4llX (%llu)", static_cast<unsigned long long>(value.ValueInt), static_cast<unsigned long long>(value.ValueInt));
                         break;
 
                     case Plip::DebugValueType::Int32Le:
                     case Plip::DebugValueType::Int32Be:
-                        ImGui::Text("0x%.8llX (%llu)", static_cast<uint64_t>(value.ValueInt), static_cast<uint64_t>(value.ValueInt));
+                        ImGui::Text("0x%.8llX (%llu)", static_cast<unsigned long long>(value.ValueInt), static_cast<unsigned long long>(value.ValueInt));
                         break;
 
                     case Plip::DebugValueType::Int64Le:
                     case Plip::DebugValueType::Int64Be:
-                        ImGui::Text("0x%.16llX (%llu)", static_cast<uint64_t>(value.ValueInt), static_cast<uint64_t>(value.ValueInt));
+                        ImGui::Text("0x%.16llX (%llu)", static_cast<unsigned long long>(value.ValueInt), static_cast<unsigned long long>(value.ValueInt));
                         break;
 
                     case Plip::DebugValueType::String:
